+++ src/main/java/pl/plajer/buildbattle/ConfigPreferences.java
@@ -31,6 +31,7 @@ import org.bukkit.Material;
 
 import pl.plajer.buildbattle.arena.impl.BaseArena;
 import pl.plajer.buildbattle.utils.Debugger;
+import pl.plajerlair.commonsbox.minecraft.compat.XMaterial;
 
 /**
  * Created by Tom on 17/08/2015.
@@ -42,6 +43,7 @@ public class ConfigPreferences {
   private List<String> endGameCommands = new ArrayList<>();
   private List<String> whitelistedCommands = new ArrayList<>();
   private List<Material> itemBlacklist = new ArrayList<>();
+  private List<Material> floorBlacklist = new ArrayList<>();
   private Map<Option, Boolean> options = new EnumMap<>(Option.class);
 
   public ConfigPreferences(Main plugin) {
@@ -105,12 +107,23 @@ public class ConfigPreferences {
     return Collections.unmodifiableList(itemBlacklist);
   }
 
+  public List<Material> getFloorBlacklist() {
+    return Collections.unmodifiableList(floorBlacklist);
+  }
+
   private void loadBlackList() {
     for (String item : plugin.getConfig().getStringList("Blacklisted-Item-Names")) {
       try {
-        itemBlacklist.add((Material.matchMaterial(item)));
+        itemBlacklist.add((XMaterial.matchXMaterial(item).get().parseMaterial()));
       } catch (IllegalArgumentException ex) {
         Bukkit.getConsoleSender().sendMessage(ChatColor.RED + "[BuildBattle] Invalid black listed item! " + item + " doesn't exist, are you sure it's properly named?");
+      }
+    }
+    for (String item : plugin.getConfig().getStringList("Blacklisted-Floor-Materials")) {
+      try {
+        floorBlacklist.add((XMaterial.matchXMaterial(item).get().parseMaterial()));
+      } catch (IllegalArgumentException ex) {
+        Bukkit.getConsoleSender().sendMessage(ChatColor.RED + "[BuildBattle] Invalid black listed material! " + item + " doesn't exist, are you sure it's properly named?");
       }
     }
   }
@@ -120,7 +133,7 @@ public class ConfigPreferences {
   }
 
   public int getTimer(TimerType type, BaseArena arena) {
-    String prefix = "Time-Manger." +arena.getArenaType().getPrefix() + ".";
+    String prefix = "Time-Manager." +arena.getArenaType().getPrefix() + ".";
     switch (type) {
       case BUILD:
         return plugin.getConfig().getInt(prefix + "Build-Time", 200);
@@ -161,7 +174,7 @@ public class ConfigPreferences {
     BOSSBAR_ENABLED("Boss-Bar-Enabled", true), BUNGEE_ENABLED("BungeeActivated", false), DATABASE_ENABLED("DatabaseActivated", false),
     INVENTORY_MANAGER_ENABLED("InventoryManager", true), BLOCK_COMMANDS_IN_GAME("Block-Commands-In-Game", true), REWARDS("Rewards-Enabled", false),
     HEADS_COMMAND("Command-Instead-Of-Head-Menu.Enabled",false), ENABLE_SHORT_COMMANDS("Enable-Short-Commands", false),
-    DISABLE_SEPARATE_CHAT("Disable-Separate-Chat", false);
+    DISABLE_SEPARATE_CHAT("Disable-Separate-Chat", false), DISABLE_PARTIES("Disable-Parties", true), ANNOUNCE_PLOTOWNER_LATER("Announce-PlotOwner-Later", false);
 
     private String path;
     private boolean def;