+++ src/main/java/pl/plajer/buildbattle/Main.java
@@ -278,7 +278,7 @@ public class Main extends JavaPlugin {
     new VoteMenuListener(this);
     new HolidayManager(this);
     BannerMenu.init(this);
-    partyHandler = new PartySupportInitializer().initialize();
+    partyHandler = new PartySupportInitializer().initialize(this);
     rewardsHandler = new RewardsFactory(this);
   }
 
@@ -319,18 +319,23 @@ public class Main extends JavaPlugin {
   private void saveAllUserStatistics() {
     for (Player player : getServer().getOnlinePlayers()) {
       User user = userManager.getUser(player);
-
-      //copy of userManager#saveStatistic but without async database call that's not allowed in onDisable method.
+      StringBuilder update = new StringBuilder(" SET ");
       for (StatsStorage.StatisticType stat : StatsStorage.StatisticType.values()) {
         if (!stat.isPersistent()) {
           continue;
         }
-        if (userManager.getDatabase() instanceof MysqlManager) {
-          ((MysqlManager) userManager.getDatabase()).getDatabase().executeUpdate("UPDATE " + ((MysqlManager) getUserManager().getDatabase()).getTableName() + " SET " + stat.getName() + "=" + user.getStat(stat) + " WHERE UUID='" + user.getPlayer().getUniqueId().toString() + "';");
-          continue;
+        if (update.toString().equalsIgnoreCase(" SET ")){
+          update.append(stat.getName()).append("=").append(user.getStat(stat));
         }
-        userManager.getDatabase().saveStatistic(user, stat);
+        update.append(", ").append(stat.getName()).append("=").append(user.getStat(stat));
       }
+      String finalUpdate = update.toString();
+      //copy of userManager#saveStatistic but without async database call that's not allowed in onDisable method.
+      if (userManager.getDatabase() instanceof MysqlManager) {
+        ((MysqlManager) userManager.getDatabase()).getDatabase().executeUpdate("UPDATE "+((MysqlManager) getUserManager().getDatabase()).getTableName()+ finalUpdate + " WHERE UUID='" + user.getPlayer().getUniqueId().toString() + "';");
+        continue;
+      }
+      userManager.getDatabase().saveAllStatistic(user);
     }
   }