+++ src/main/java/pl/plajer/buildbattle/commands/arguments/admin/ListArenasArgument.java
@@ -50,7 +50,7 @@ public class ListArenasArgument {
         }
         if (i == 0) {
           sender.sendMessage(registry.getPlugin().getChatManager().colorMessage("Commands.Admin-Commands.List-Command.No-Arenas"));
-          sender.sendMessage(registry.getPlugin().getChatManager().colorRawMessage("&e&lTIP: &7You can get free maps with configs at our wiki! Just head to https://wiki.plajer.xyz/minecraft/buildbattle/free_maps.php"));
+          sender.sendMessage(registry.getPlugin().getChatManager().colorRawMessage("&e&lTIP: &7You can get free maps with configs at our wiki! Just head to https://wiki.plugily.xyz/minecraft/buildbattle/free_maps.php"));
         }
       }
     });