+++ src/main/java/pl/plajer/buildbattle/events/QuitEvents.java
@@ -58,9 +58,7 @@ public class QuitEvents implements Listener {
   @EventHandler
   public void onQuitSaveStats(PlayerQuitEvent e) {
     User user = plugin.getUserManager().getUser(e.getPlayer());
-    for (StatsStorage.StatisticType stat : StatsStorage.StatisticType.values()) {
-      plugin.getUserManager().saveStatistic(user, stat);
-    }
+    plugin.getUserManager().saveAllStatistic(user);
     plugin.getUserManager().removeUser(user);
   }