+++ src/main/java/pl/plajer/buildbattle/handlers/language/LanguageMigrator.java
@@ -40,8 +40,8 @@ import pl.plajerlair.commonsbox.minecraft.migrator.MigratorUtils;
  */
 public class LanguageMigrator {
 
-  public static final int LANGUAGE_FILE_VERSION = 15;
-  public static final int CONFIG_FILE_VERSION = 10;
+  public static final int LANGUAGE_FILE_VERSION = 16;
+  public static final int CONFIG_FILE_VERSION = 13;
   private List<String> migratable = Arrays.asList("bungee", "config", "language", "mysql");
   private Main plugin;
 
@@ -198,6 +198,24 @@ public class LanguageMigrator {
                   "# It will disable the separated chat, for example\r\n" +
                   "Disable-Separate-Chat: false\r\n");
           break;
+        case 10:
+          MigratorUtils.addNewLines(file, "\r\n# Blacklisted floor materials, you can't use them for the floor.\r\n" +
+                  "Blacklisted-Floor-Materials:\r\n" +
+                  "  - MOB_SPAWNER\r\n" +
+                  "  - TNT\r\n" +
+                  "  - AIR\r\n");
+          break;
+        case 11:
+          MigratorUtils.addNewLines(file, "\r\n" +
+                  "#Disable Party features of external party plugins (such as PAF, Parties ...)\r\n" +
+                  "Disable-Parties: true\r\n");
+          break;
+        case 12:
+          MigratorUtils.addNewLines(file, "\r\n" +
+                  "#Announce the plot owner after voting stage\r\n" +
+                  "#default false - plot owner will be announced at the beginning\r\n" +
+                  "Announce-PlotOwner-Later: false \r\n");
+          break;
       }
     }
     Bukkit.getConsoleSender().sendMessage(ChatColor.GREEN + "[BuildBattle] [System notify] Config updated, no comments were removed :)");
@@ -281,16 +299,16 @@ public class LanguageMigrator {
               "    Build-Time: \"&fTime: &e%FORMATTED_TIME_LEFT%\"\r\n    Starts-In: \"&fStarts In: &e%FORMATTED_TIME_LEFT%\"");
           MigratorUtils.insertAfterLine(file, "Content:", "    Playing-States:\r\n      Classic:\r\n        - \"&fTime Left: &e%FORMATTED_TIME_LEFT%\"\r\n        - \"\"\r\n" +
               "        - \"&fPlayers: &e%PLAYERS%/%MAX_PLAYERS%\"\r\n        - \"\"\r\n        - \"&fTheme: &e%THEME%\"\r\n        - \"\"\r\n        - \"&fArena: &e%ARENA_ID%\"\r\n" +
-              "        - \"\"\r\n        - \"&ewww.plajer.xyz\"\r\n      Teams:\r\n        - \"&7Teams Mode\"\r\n        - \"&fTime Left: &e%FORMATTED_TIME_LEFT%\"\r\n        - \"\"\r\n" +
+              "        - \"\"\r\n        - \"&ewww.plugily.xyz\"\r\n      Teams:\r\n        - \"&7Teams Mode\"\r\n        - \"&fTime Left: &e%FORMATTED_TIME_LEFT%\"\r\n        - \"\"\r\n" +
               "        - \"&fTheme: &e%THEME%\"\r\n        - \"\"\r\n        - \"&fArena: &e%ARENA_ID%\"\r\n        - \"\"\r\n        - \"&fTeammate:\"\r\n        - \"&e%TEAMMATE%\"\r\n" +
-              "        - \"\"\r\n        - \"&ewww.plajer.xyz\"\r\n      Guess-The-Build:\r\n        - \"&7Guess The Build Mode\"\r\n        - \"&fBuilder:\"\r\n        - \"&7%BUILDER%\"\r\n" +
+              "        - \"\"\r\n        - \"&ewww.plugily.xyz\"\r\n      Guess-The-Build:\r\n        - \"&7Guess The Build Mode\"\r\n        - \"&fBuilder:\"\r\n        - \"&7%BUILDER%\"\r\n" +
               "        - \"\"\r\n        - \"&e&lLeaders:\"\r\n        - \"&6%1%&e: %1_PTS%\"\r\n        - \"&7%2%&e: %2_PTS%\"\r\n        - \"&7%3%&e: %3_PTS%\"\r\n        - \"\"\r\n" +
-              "        - \"%CURRENT_TIMER%\"\r\n        - \"\"\r\n        - \"&fTheme:\"\r\n        - \"&c%THEME%\"\r\n        - \"\"\r\n        - \"&ewww.plajer.xyz\"");
-          MigratorUtils.insertAfterLine(file, "Content:", "    Ending-States:\r\n      Classic:\r\n        - \"&e&lGAME ENDED\"\r\n        - \"\"\r\n        - \"&ewww.plajer.xyz\"\r\n" +
-              "      Teams:\r\n        - \"&7Teams Mode\"\r\n        - \"&e&lGAME ENDED\"\r\n        - \"\"\r\n        - \"&ewww.plajer.xyz\"\r\n      Guess-The-Build:\r\n" +
+              "        - \"%CURRENT_TIMER%\"\r\n        - \"\"\r\n        - \"&fTheme:\"\r\n        - \"&c%THEME%\"\r\n        - \"\"\r\n        - \"&ewww.plugily.xyz\"");
+          MigratorUtils.insertAfterLine(file, "Content:", "    Ending-States:\r\n      Classic:\r\n        - \"&e&lGAME ENDED\"\r\n        - \"\"\r\n        - \"&ewww.plugily.xyz\"\r\n" +
+              "      Teams:\r\n        - \"&7Teams Mode\"\r\n        - \"&e&lGAME ENDED\"\r\n        - \"\"\r\n        - \"&ewww.plugily.xyz\"\r\n      Guess-The-Build:\r\n" +
               "        - \"&7Guess The Build Mode\"\r\n        - \"\"\r\n        - \"&e1. &f%1%: &e%1_PTS%\"\r\n        - \"&e2. &f%2%: &e%2_PTS%\"\r\n        - \"&e3. &f%3%: &e%3_PTS%\"\r\n" +
               "        - \"&e4. &f%4%: &e%4_PTS%\"\r\n        - \"&e5. &f%5%: &e%5_PTS%\"\r\n        - \"&e6. &f%6%: &e%6_PTS%\"\r\n        - \"&e7. &f%7%: &e%7_PTS%\"\r\n" +
-              "        - \"&e8. &f%8%: &e%8_PTS%\"\r\n        - \"&e9. &f%9%: &e%9_PTS%\"\r\n        - \"&e10. &f%10%: &e%10_PTS%\"\r\n        - \"\"\r\n        - \"&ewww.plajer.xyz\"");
+              "        - \"&e8. &f%8%: &e%8_PTS%\"\r\n        - \"&e9. &f%9%: &e%9_PTS%\"\r\n        - \"&e10. &f%10%: &e%10_PTS%\"\r\n        - \"\"\r\n        - \"&ewww.plugily.xyz\"");
           MigratorUtils.insertAfterLine(file, "No-Theme-Yet", "  Guess-The-Build:\r\n    Current-Builder: \"&eBuilder: &7%BUILDER%\"\r\n" +
               "    Current-Round: \"&eRound: &7%ROUND%/%MAXPLAYERS%\"\r\n    Theme-Is-Long: \"&eThe theme is &7%NUM% &echaracters long\"\r\n    Theme-Is-Name: \"&eThe theme is a &e%THEME%\"");
           MigratorUtils.insertAfterLine(file, "Winner-Title:", "      Summary:\r\n" +
@@ -399,6 +417,10 @@ public class LanguageMigrator {
           break;
         case 14:
           MigratorUtils.insertAfterLine(file, "Theme-Was-Subtitle:", "    Theme-Guessed: \"&eAll players guessed the theme!\"");
+          break;
+        case 15:
+          MigratorUtils.insertAfterLine(file, "Voting-Messages:", "      Voted-For-Player-Plot: \"&7The plot owner was &e&l%PLAYER%&7!\"");
+          MigratorUtils.insertAfterLine(file, "Voting-Messages:", "      Vote-For-Next-Plot: \"&7You can now vote for this plot!\"");
           break;
       }
       version++;